setting message for gelf
Instead of allowing the message to get set to the full_message, we're
setting message to the message that was logged. This fixes the problem
where we have trace turned on, which sets the message to what the full
message is.

This will make things easier to read/understand in cases where Graylog2
is used to display log messages.